refactor: move resources to builder
