--- rust/operator-binary/src/config/mod.rs
@@ -1,3 +1,0 @@
-pub mod command;
-pub mod jvm;
-pub mod node_id_hasher;
