+++ rust/operator-binary/src/controller/node_id_hasher.rs
