--- rust/operator-binary/src/operations/mod.rs
@@ -1,2 +1,0 @@
-pub mod graceful_shutdown;
-pub mod pdb;
