--- rust/operator-binary/src/resource/mod.rs
@@ -1,3 +1,0 @@
-pub mod listener;
-pub mod service;
-pub mod statefulset;
